Revert "errors: Support multierr and pkg/errors (#460)"
This reverts commit acea9071443b209a0df36bbd68adf06708fe9f82.
