+++ zapcore/field_test.go
@@ -23,12 +23,14 @@ package zapcore_test
 import (
 	"errors"
 	"fmt"
+	"io"
 	"math"
 	"testing"
 	"time"
 
 	"go.uber.org/zap"
 
+	richErrors "github.com/pkg/errors"
 	"github.com/stretchr/testify/assert"
 
 	. "go.uber.org/zap/zapcore"
@@ -38,6 +40,18 @@ type users int
 
 func (u users) String() string {
 	return fmt.Sprintf("%d users", int(u))
+}
+
+func (u users) Error() string {
+	return fmt.Sprintf("%d too many users", int(u))
+}
+
+func (u users) Format(s fmt.State, verb rune) {
+	// Implement fmt.Formatter, but don't add any information beyond the basic
+	// Error method.
+	if verb == 'v' && s.Flag('+') {
+		io.WriteString(s, u.Error())
+	}
 }
 
 func (u users) MarshalLogObject(enc ObjectEncoder) error {
@@ -116,6 +130,7 @@ func TestFields(t *testing.T) {
 		{t: ReflectType, iface: users(2), want: users(2)},
 		{t: NamespaceType, want: map[string]interface{}{}},
 		{t: StringerType, iface: users(2), want: "2 users"},
+		{t: ErrorType, iface: users(2), want: "2 too many users"},
 		{t: SkipType, want: interface{}(nil)},
 	}
 
@@ -130,6 +145,24 @@ func TestFields(t *testing.T) {
 
 		assert.True(t, f.Equals(f), "Field does not equal itself")
 	}
+}
+
+func TestRichErrorSupport(t *testing.T) {
+	f := Field{
+		Type:      ErrorType,
+		Interface: richErrors.WithMessage(richErrors.New("egad"), "failed"),
+		Key:       "k",
+	}
+	enc := NewMapObjectEncoder()
+	f.AddTo(enc)
+	assert.Equal(t, "failed: egad", enc.Fields["k"], "Unexpected basic error message.")
+
+	serialized := enc.Fields["kVerbose"]
+	// Don't assert the exact format used by a third-party package, but ensure
+	// that some critical elements are present.
+	assert.Regexp(t, `egad`, serialized, "Expected original error message to be present.")
+	assert.Regexp(t, `failed`, serialized, "Expected error annotation to be present.")
+	assert.Regexp(t, `TestRichErrorSupport`, serialized, "Expected calling function to be present in stacktrace.")
 }
 
 func TestEquals(t *testing.T) {
